# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "DB project server.",
        "title": "DB project API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:5000",
    "basePath": "/api",
    "paths": {
...
        "/post/{id}/details": {
            "get": {
                "description": "Gets post by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Post"
                ],
                "summary": "Gets post by id",
                "operationId": "GetPostById",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of post",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "related",
                        "name": "related",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.PostFull"
                        }
                    },
                    "404": {
                        "description": "Not found - Requested entity is not found in database",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Request is valid but operation failed at server side",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            },
            "post": {
                "description": "Edits post by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Post"
                ],
                "summary": "Editss post by id",
                "operationId": "UpdatePost",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of post",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "PostUpdate params",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.PostUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.PostFull"
                        }
                    },
                    "404": {
                        "description": "Not found - Requested entity is not found in database",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Request is valid but operation failed at server side",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
...
        "/thread/{slug_or_id}/posts": {
            "get": {
                "description": "GetThreadPosts",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Thread"
                ],
                "summary": "GetThreadPosts",
                "operationId": "GetThreadPosts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "slug or id of thread",
                        "name": "slug_or_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "since",
                        "name": "since",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "sort",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "desc",
                        "name": "desc",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Thread"
                            }
                        }
                    },
                    "404": {
                        "description": "Not found - Requested entity is not found in database",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Request is valid but operation failed at server side",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
...
        "/user/{nickname}/profile": {
            "get": {
                "description": "Gets Users profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Gets Users profile",
                "operationId": "GetProfile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "nickname of user",
                        "name": "nickname",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    },
                    "404": {
                        "description": "Not found - Requested entity is not found in database",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Request is valid but operation failed at server side",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            },
            "post": {
                "description": "Changes Users profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Changes Users profile",
                "operationId": "PostProfile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "nickname of user",
                        "name": "nickname",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User params",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "404": {
                        "description": "Not found - Requested entity is not found in database",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    },
                    "409": {
                        "description": "Conflict - User already exists",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Request is valid but operation failed at server side",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Error": {
            "type": "object",
            "properties": {
                "error": {}
            }
        },
        "model.Forum": {
            "type": "object",
            "properties": {
                "posts": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                },
                "threads": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "user": {
                    "type": "string"
                }
            }
        },
...
        "model.Post": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "created": {
                    "type": "string"
                },
                "forum": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isEdited": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                },
                "parent": {
                    "type": "integer"
                },
                "thread": {
                    "type": "integer"
                }
            }
        },
        "model.PostFull": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/model.User"
                },
                "forum": {
                    "$ref": "#/definitions/model.Forum"
                },
                "post": {
                    "$ref": "#/definitions/model.Post"
                },
                "thread": {
                    "$ref": "#/definitions/model.Thread"
                }
            }
        },
        "model.PostUpdate": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "model.Response": {
            "type": "object",
            "properties": {
                "body": {}
            }
        },
...
        "model.Thread": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "created": {
                    "type": "string"
                },
                "forum": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "votes": {
                    "type": "integer"
                }
            }
        },
...
        "model.User": {
            "type": "object",
            "properties": {
                "about": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "fullname": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                }
            }
        },
...
    }
}